* Updated the version of items.otb to 2 * All attributes that can be changed without any problems has now been moved from the items.otb to items.xml for easier manipulation (users upgrading to this revision will need to update their items.otb and items.xml) * Fixed some compiling warnings in MSVC * Fixed a bug in Combat::CombatTypeToCondition

+++ otserv/trunk/chat.cpp
@@ -193,7 +193,7 @@ ChatChannel* Chat::createChannel(Player* player, uint16_t channelId)
 		}
 		
 		//find a free private channel slot
-		for(uint32_t i = 100; i < 10000; ++i){
+		for(uint16_t i = 100; i < 10000; ++i){
 			if(m_privateChannels.find(i) == m_privateChannels.end()){
 				PrivateChatChannel* newChannel = new PrivateChatChannel(i, player->getName() + "'s Channel");
 				if(!newChannel)
